Prep for move to 0.8
Edits as per work group discussion to get ready for 0.8:

1. Prune/remove all material not directly related to the standard. That
includes history, justifications and other material that should be in a
wiki or other collateral.

2. Brought in the API field specification from the Consent Generator

3. Converted the 5 sections of the spec to tabular format

4. Modified the top o the document to match the UMA spec (since that is
a spec that is already accepted, I thought that the way of presenting
the document information the abstract and the IP/copyright were valid.
Allowed me to add editors and contributors.

5. Wordsmithing (grammar and typos)

6. Because the simplified format is more clear, it suggests some
necessary changes. For example “Method of collection” in the API
conflated the type of consent (expressed vs implied) with how that
information was collected. Made minimal changes/additions to address
this and similar issues.

To do:

Work group discussion of format (i.e. are tables more clear, did I
remove too much detritus, IP header, etc)

Review of fields in tables to simplify and change, remove or add to get
to minimum viable.

Discuss who/how/if to update generator to align with version 0.8 of
spec when accepted.